added intro screen
we still need to write the string for how to play the game or if you
want to make the intro screen a diff way then just let me know and
change it. This is just an idea. text me after you check this out so i
know what you think